Convert FloorItems
