+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/item/ItemPickup.kts
@@ -9,18 +9,18 @@ import world.gregs.voidps.engine.entity.character.player.chat.inventoryFull
 import world.gregs.voidps.engine.entity.character.setAnimation
 import world.gregs.voidps.engine.entity.character.turn
 import world.gregs.voidps.engine.entity.item.floor.FloorItemOption
-import world.gregs.voidps.engine.entity.item.floor.FloorItems
+import world.gregs.voidps.engine.entity.item.floor.FloorItemStorage
 import world.gregs.voidps.engine.event.on
 import world.gregs.voidps.engine.inject
 import world.gregs.voidps.world.interact.entity.sound.playSound
 
-val items: FloorItems by inject()
+val floorItems: FloorItemStorage by inject()
 val logger = InlineLogger()
 
 on<FloorItemOption>({ operate && option == "Take" }) { player: Player ->
     if (player.inventory.isFull() && (!player.inventory.stackable(item.id) || !player.inventory.contains(item.id))) {
         player.inventoryFull()
-    } else if (items.remove(item)) {
+    } else if (floorItems.remove(item)) {
         player.inventory.add(item.id, item.amount)
         when (player.inventory.transaction.error) {
             TransactionError.None -> {
